+++ WorkloadManagementSystem/Client/PilotsLoggingClient.py
@@ -1,43 +1,40 @@
-""" Class that contains client access to the PilotsLogging handler. """
-
+""" Class that contains client access to the PilotAgents handler. """
+########################################################################
+# $Id$
+# $HeadURL$
+########################################################################
 __RCSID__ = "$Id$"
 
-from DIRAC.Core.Base.Client import Client
+from DIRAC.Core.Base.Client                         import Client
 
 class PilotsLoggingClient( Client ):
-  """Implementation of interface of Pilots Logging service. Client class should be used to communicate
-  with PilotsLogging Service
-  """
 
   def __init__( self, **kwargs ):
     Client.__init__( self, **kwargs )
     self.setServer( 'WorkloadManagement/PilotsLogging' )
     self.pilotsLoggingHandler = self._getRPC()
 
-  def addPilotsLogging( self, pilotRef, status, minorStatus, timeStamp, source ):
-    """
-    Add new Pilots Logging entry
-    :param pilotRef: Pilot reference
-    :param status: Pilot status
-    :param minorStatus: Additional status information
-    :param timeStamp: Date and time of status event
-    :param source: Source of statu information
-    """
+  def addPilotsLogging( self, pilotUUID, status, minorStatus, timeStamp, source, pilotID ):
 
-    return self.pilotsLoggingHandler.addPilotsLogging(pilotRef, status, minorStatus, timeStamp, source)
+    resp = self.pilotsLoggingHandler.addPilotsUUID( pilotUUID )
+    if not resp['OK']:
+      return resp
 
-  def deletePilotsLogging( self, pilotRef ):
-    """
-    Delete all Logging entries for Pilot
-    :param pilotRef: Pilot reference
-    """
+    if pilotID:
+      resp = self.pilotsLoggingHandler.setPilotsUUIDtoIDMapping( pilotUUID, pilotID )
+      if not resp['OK']:
+        return resp
 
-    return self.pilotsLoggingHandler.detelePilotsLogging( pilotRef )
+    return self.pilotsLoggingHandler.addPilotsLogging(pilotUUID, status, minorStatus, timeStamp, source)
 
-  def getPilotsLogging( self, pilotRef ):
-    """
-    Get all Logging entries for Pilot
-    :param pilotRef: Pilot reference
-    """
+  def setPilotsUUIDtoIDMapping( self, pilotUUID, pilotID ):
 
-    return self.pilotsLoggingHandler.getPilotsLogging( pilotRef )
+    return self.pilotsLoggingHandler.setPilotsUUIDtoIDMapping( pilotUUID, pilotID )
+
+  def deletePilotsLogging( self, pilotID ):
+
+    return self.pilotsLoggingHandler.detelePilotsLogging( pilotID )
+
+  def getPilotsLogging( self, pilotID ):
+
+    return self.pilotsLoggingHandler.getPilotsLogging( pilotID )
