+++ WorkloadManagementSystem/Service/PilotsLoggingHandler.py
@@ -14,53 +14,38 @@ __RCSID__ = "$Id: $"
 
 from DIRAC.WorkloadManagementSystem.DB.PilotsLoggingDB import PilotsLoggingDB
 
+import types
 from DIRAC.Core.DISET.RequestHandler import RequestHandler
-from DIRAC import S_OK
+from DIRAC import gLogger, S_OK, S_ERROR
+from DIRAC.Core.Utilities import Time
+# from DIRAC.WorkloadManagementSystem.DB import PilotAgentsDB
 
 class PilotsLoggingHandler( RequestHandler ):
-  """Server side functions for Pilots Logging service"""
 
   @classmethod
   def initializeHandler( cls, serviceInfo ):
-    """Initialization of Pilots Logging service
-    """
     return S_OK()
 
   def initialize(self):
-    """Initialization of Pilots Logging service
-    """
     self.pilotsLogging = PilotsLoggingDB()
 
-  types_addPilotsLogging = [ basestring, basestring, basestring, float, basestring ]
-  def export_addPilotsLogging( self, pilotRef, status, minorStatus, timeStamp, source ):
-    """
-    Add new Pilots Logging entry
-    :param pilotRef: Pilot reference
-    :param status: Pilot status
-    :param minorStatus: Additional status information
-    :param timeStamp: Date and time of status event
-    :param source: Source of statu information
-    """
+  auth_addPilotsLogging = [ 'all' ]
+  types_addPilotsLogging = [ types.StringType, types.StringType, types.StringType, types.FloatType, types.StringType ]
+  def export_addPilotsLogging( self, pilotUUID, status, minorStatus, timeStamp, source ):
 
-    return self.pilotsLogging.addPilotsLogging( pilotRef, status, minorStatus, timeStamp, source )
+    return self.pilotsLogging.addPilotsLogging( pilotUUID, status, minorStatus, timeStamp, source )
 
-  types_getPilotsLogging = [ basestring ]
-  def export_getPilotsLogging( self, pilotRef ):
-    """
-    Get all Logging entries for Pilot
-    :param pilotRef: Pilot reference
-    """
+  auth_getPilotsLogging = [ 'all' ]
+  types_getPilotsLogging = [ types.IntType ]
+  def export_getPilotsLogging( self, pilotID ):
 
-    return self.pilotsLogging.getPilotsLogging( pilotRef )
+    return self.pilotsLogging.getPilotsLogging( pilotID )
 
-  types_deletePilotsLogging = [ [basestring,  list] ]
-  def export_deletePilotsLogging( self, pilotRef ):
-    """
-    Delete all Logging entries for Pilot
-    :param pilotRef: Pilot reference
-    """
+  auth_setPilotsUUIDtoIDMapping = [ 'all' ]
+  types_setPilotsUUIDtoIDMapping = [ types.StringType, types.IntType ]
+  def export_setPilotsUUIDtoIDMapping( self, pilotUUID, pilotID ):
 
-    return self.pilotsLogging.deletePilotsLogging( pilotRef )
+    return self.pilotsLogging.setPilotsUUIDtoIDMapping( pilotUUID, pilotID )
 
   types_addPilotsUUID = [ basestring ]
   auth_addPilotsUUID = [ 'Operator' ]
